storage/directory.py: Move _FileType to CasBasedDirectory
This is now entirely an internal implementation detail of how
CasBasedDirectory handles it's internal IndexEntry objects.
